[REFACTOR] youtube 디렉토리 정리
