+++ src/main/java/channeling/be/domain/channel/application/ChannelServiceImpl.java
@@ -9,8 +9,8 @@ import channeling.be.infrastructure.kafka.producer.KafkaProducerService;
 import channeling.be.infrastructure.kafka.dto.KafkaVideoSyncDto;
 import channeling.be.infrastructure.redis.RedisUtil;
 import channeling.be.infrastructure.youtube.application.YoutubeService;
-import channeling.be.infrastructure.youtube.dto.YoutubeDto;
-import channeling.be.infrastructure.youtube.dto.res.YoutubeChannelResDTO;
+import channeling.be.infrastructure.youtube.presentation.YoutubeDto;
+import channeling.be.infrastructure.youtube.res.YoutubeChannelResDTO;
 import channeling.be.response.exception.handler.ChannelHandler;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
