+++ src/main/java/channeling/be/infrastructure/kafka/dto/KafkaVideoSyncDto.java
@@ -1,7 +1,7 @@
 package channeling.be.infrastructure.kafka.dto;
 
 import channeling.be.domain.channel.domain.Channel;
-import channeling.be.infrastructure.youtube.dto.res.YoutubeChannelResDTO;
+import channeling.be.infrastructure.youtube.res.YoutubeChannelResDTO;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
