+++ src/main/java/channeling/be/infrastructure/youtube/application/YoutubeServiceImpl.java
@@ -3,11 +3,11 @@ package channeling.be.infrastructure.youtube.application;
 import channeling.be.domain.channel.application.ChannelServiceImpl;
 import channeling.be.domain.channel.domain.Channel;
 import channeling.be.domain.video.application.VideoService;
-import channeling.be.infrastructure.youtube.YoutubeConvertor;
-import channeling.be.infrastructure.youtube.dto.YoutubeDto;
-import channeling.be.infrastructure.youtube.dto.model.YoutubeVideoListResDTO;
-import channeling.be.infrastructure.youtube.dto.res.YoutubeChannelResDTO;
-import channeling.be.infrastructure.youtube.dto.res.YoutubePlayListResDTO;
+import channeling.be.infrastructure.youtube.presentation.YoutubeConvertor;
+import channeling.be.infrastructure.youtube.presentation.YoutubeDto;
+import channeling.be.infrastructure.youtube.res.YoutubeVideoListResDTO;
+import channeling.be.infrastructure.youtube.res.YoutubeChannelResDTO;
+import channeling.be.infrastructure.youtube.res.YoutubePlayListResDTO;
 import channeling.be.response.code.status.ErrorStatus;
 import channeling.be.response.exception.handler.YoutubeHandler;
 import lombok.RequiredArgsConstructor;
